requirements.txt
discord
pysqlite3
rich
pyyaml
